# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Piazza API",
        "description": "API documentation for Piazza.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/team-members": {
            "get": {
                "tags": [
                    "Team Members"
                ],
                "summary": "Get all team members",
                "operationId": "b34e8010b0fe4331a0934852350422cf",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filter team members by role",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (active, inactive)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name or bio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of team members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TeamMember"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Team Members"
                ],
                "summary": "Create a new team member",
                "operationId": "ebceff7b8479b39a15c4274fb60638fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamMember"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Team member created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/team-members/{id}": {
            "get": {
                "tags": [
                    "Team Members"
                ],
                "summary": "Get a single team member",
                "operationId": "cd8cadc1715392bfc4f329323f93dd72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team member details"
                    },
                    "404": {
                        "description": "Team member not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Team Members"
                ],
                "summary": "Update a team member",
                "operationId": "d6424fd97eae26513aa3e1b7873634a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamMember"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team member updated successfully"
                    },
                    "404": {
                        "description": "Team member not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/team-members/{id}": {
            "delete": {
                "tags": [
                    "Team Members"
                ],
                "summary": "Delete a team member",
                "operationId": "f1bc65ac23b573e2cb7735a83630c23c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Team member deleted"
                    },
                    "404": {
                        "description": "Team member not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TeamMember": {
                "required": [
                    "name",
                    "role"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "role": {
                        "type": "string",
                        "example": "Project Manager"
                    },
                    "profile_image": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "example": "Experienced project manager with a focus on agile methodologies.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}